+++ Data Structures/Mo's Algorithm/Mo's algorithm from geeksforgeeks.cpp
